requirements.txt
flask==3.0.3
mysql-connector-python==8.3.0
python-dotenv==1.0.1